TCP: Fix the Flow Counter statistics for deinterlaced conversations
The conversations dialog always give a wrong number of TCP Flows
when any deinterlacing key is applied.

Implement a Map to track the relationship between TCP Stream IDs
and Conversations, then the conversation, the conversation table,
and the stream ID which are indirectly correlated,
can be easily linked together for this functionality.